% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artful.R
\docType{package}
\name{artful-package}
\alias{artful}
\alias{artful-package}
\title{artful: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Page \email{mike@cynkra.com}

}
\keyword{internal}
