% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{create_default_labels}
\alias{create_default_labels}
\title{Create default labels from stat names}
\usage{
create_default_labels(stat_names)
}
\description{
Create default labels from stat names
}
\examples{
\dontrun{
create_default_labels("hr")
create_default_labels("No match found")
}
}
\keyword{internal}
