% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{create_missing_result}
\alias{create_missing_result}
\title{Create result for missing values}
\usage{
create_missing_result(stat_value)
}
\arguments{
\item{stat_value}{Original value}
}
\value{
Tibble with "missing" classification
}
\description{
Create result for missing values
}
