% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{create_unmatched_result}
\alias{create_unmatched_result}
\title{Create result for unmatched values}
\usage{
create_unmatched_result(stat_value)
}
\arguments{
\item{stat_value}{Original value}
}
\value{
Tibble with "other" classification
}
\description{
Create result for unmatched values
}
