% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\docType{data}
\name{.stateful_patterns}
\alias{.stateful_patterns}
\title{Global Pattern Registry Environment}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.stateful_patterns
}
\description{
Global Pattern Registry Environment
}
\keyword{internal}
