% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get_default_patterns}
\alias{get_default_patterns}
\title{Get default statistical patterns}
\usage{
get_default_patterns()
}
\value{
Named list of pattern definitions
}
\description{
Get default statistical patterns
}
