% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get_stat_patterns}
\alias{get_stat_patterns}
\title{Get statistical patterns}
\usage{
get_stat_patterns()
}
\value{
Named list of statistical pattern definitions
}
\description{
Returns the current global statistical patterns.
}
