% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{html_to_dataframe}
\alias{html_to_dataframe}
\title{Convert a string of HTML into a dataframe}
\usage{
html_to_dataframe(html)
}
\arguments{
\item{html}{A string, HTML generated by calling \url{rtf_to_html}.}
}
\value{
A data frame containing an R native version of the original RTF
table.
}
\description{
Convert the HTML generated by calling \url{rtf_to_html} into a data frame.
}
\keyword{internal}
