% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{indentation_to_variables}
\alias{indentation_to_variables}
\title{Convert indentation to variable cols}
\usage{
indentation_to_variables(data)
}
\description{
Unlike `strip_indentation(), this function is designed to be called after a
dataframe has been generated from an RTF instead of during the generation.
In effect this function pivots a data frame wider, using indentation to infer
how the pivoting should happen.
}
\examples{
\dontrun{
indentation_to_variables(rtf_to_df(rtf_example(13)))
}
}
