% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{manage_exceptions}
\alias{manage_exceptions}
\title{Manage table expcetions}
\usage{
manage_exceptions(data)
}
\description{
Several tables do not align with the general ruleset and must be treated as
exceptions that get special processing to re-align them.
}
\keyword{internal}
