% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{nbsp_to_spaces}
\alias{nbsp_to_spaces}
\title{Replace non-breaking spaces with spaces}
\usage{
nbsp_to_spaces(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=strip_pagination]{strip_pagination()}} containing
indentation in the first column, represented by " " values in each cell.}
}
\value{
A data frame with all isntances of " " replaced by empty spaces
(e.g., " ").
}
\description{
Convert indentation identified by " " (create via
\code{rtf_spaces_to_nbsp()}) with empty spaces (e.g., " ").
}
\keyword{internal}
