% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pivot_group}
\alias{pivot_group}
\title{Pivot a grouping variable longer}
\usage{
pivot_group(data)
}
\arguments{
\item{data}{A data frame with the grouping variable in a wide format.}
}
\value{
A data frame with the grouping variable in a long format.
}
\description{
The grouping variables in the final ARD are normally presented in a wide
format in the input RTF table. This functions pivots the grouping levels into
a long format.
}
\keyword{internal}
