% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pivot_indentation}
\alias{pivot_indentation}
\title{Pivot an indented data frame wider}
\usage{
pivot_indentation(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=split_data]{split_data()}}.}
}
\value{
A data frame with indentation removed and placed into new columns.
}
\description{
This function pivots indentation in a data frame into a wide format.
Indentation is determined by empty strings in the second column with the
number of consectuive empty strings in the first n rows representing n
indentation levels. For two levels of indentation, the first level of
indentation appears only once, at the top of the data frame. The second level
of indentation may appear n times through the data frame. This function only
supports indentation levels up to two. A more general solution to support
2+ levels is sketched out in \code{inst/scripts/pivot-n.R}. For now, the simpler
solution below is preferred.
}
\keyword{internal}
