% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{rtf_to_df}
\alias{rtf_to_df}
\title{Convert an RTF Table into a data frame}
\usage{
rtf_to_df(file)
}
\arguments{
\item{file}{A string, the path to the input .rtf file.}
}
\value{
an R data frame.
}
\description{
This function converts a table in RTF format into a data frame in R. The
function aims to return a copy of the original RTF table with as little data
preprocessing performed as possible (i.e., only essential steps such as
removing headers, footers, empty columns, etc., are performed). This function
does not attempt to produce an ARD. Instead, this step is off-loaded to other
functions.
}
\examples{
\dontrun{
rtf_to_df(rtf_example(1))
purrr::map(1:21, ~ rtf_to_df(rtf_example(.x)))
}
}
