% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{rtf_to_html}
\alias{rtf_to_html}
\title{Convert an RTF file to HTML using Pandoc}
\usage{
rtf_to_html(file)
}
\arguments{
\item{file}{A string, the path to the input .rtf file.}
}
\value{
Character vector of HTML.
}
\description{
This function takes the path to an input RTF file and uses the Pandoc
command-line tool to convert it to HTML.
}
\keyword{internal}
