% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{separate_bign}
\alias{separate_bign}
\title{Separate the Big N stat}
\usage{
separate_bign(data)
}
\arguments{
\item{data}{A data frame containing a column called "group1_level" which
contains the Big N statistic merged as part of the text.}
}
\value{
A data frame with the Big N stat separated and placed at the top.
}
\description{
Extract the Big N stat and place it at the top of a table.
}
\keyword{internal}
