% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{separate_merges}
\alias{separate_merges}
\title{Separate a merged columns into multiple columns}
\usage{
separate_merges(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame where merged columns have been separated.
}
\description{
This function separates columns containing merged values (e.g., where a count
and percentage are merged into a single string such as "34 (.56)") into
multiple columns inside a data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.
}
\keyword{internal}
