% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{set_colnames}
\alias{set_colnames}
\title{Set column names}
\usage{
set_colnames(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame with the repeat column names removed.
}
\description{
This function sets columns names inside a data frame generated by calling
\code{\link[=rtf_to_html]{rtf_to_html()}}. It assumes the first row contains the actual column names.
}
\keyword{internal}
