% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{set_stat_patterns}
\alias{set_stat_patterns}
\title{Set statistical patterns}
\usage{
set_stat_patterns(patterns)
}
\arguments{
\item{patterns}{Named list of pattern definitions}
}
\description{
Set statistical patterns
}
