% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{split_data}
\alias{split_data}
\title{Split a data frame into multiple tables by indentation}
\usage{
split_data(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=strip_pagination]{strip_pagination()}}.}
}
\value{
A list of data frames, each containing the same repeated levels of
indentation (represented by empty strings in the second column).
}
\description{
This function splits a data frame into a list of data frames where each new
data frame contains the same repeated levels of indentation. Indentation is
determined by empty strings in the second column with the number of
consectuive empty strings in the first n rows representing n indentation
levels.
}
\keyword{internal}
