% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{strip_empty_rows}
\alias{strip_empty_rows}
\title{Remove empty rows}
\usage{
strip_empty_rows(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame with empty rows removed.
}
\description{
This function removes empty rows (where all cells are empty strings) inside a
data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.
}
\keyword{internal}
