% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{strip_footer}
\alias{strip_footer}
\title{Remove the footer}
\usage{
strip_footer(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame with the footer removed.
}
\description{
This function removes the rows representing a footer inside a data frame
generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.
}
\keyword{internal}
