% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{strip_header}
\alias{strip_header}
\title{Remove the header}
\usage{
strip_header(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame with the header removed.
}
\description{
This function removes the rows representing a header (e.g., titles and
subtitles) inside a data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.
}
\keyword{internal}
