% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{strip_indentation}
\alias{strip_indentation}
\title{Strip indentation}
\usage{
strip_indentation(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=strip_pagination]{strip_pagination()}} containing
indentation in the first column, represented by " " values in each cell.}
}
\value{
A data frame where the indented first column has been parsed into
separate \verb{variable_*} columns in accordance with the ARD standard.
}
\description{
Convert indentation identified by " " values into a series of
\verb{variable_*} columns.
}
\keyword{internal}
