% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{strip_repeat_colnames}
\alias{strip_repeat_colnames}
\title{Remove repeat column names}
\usage{
strip_repeat_colnames(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame with the repeat column names removed.
}
\description{
This function removes columns names that repeat across multiple rows (caused
by pagination in the original RTF files) inside a data frame generated by
calling \code{\link[=rtf_to_html]{rtf_to_html()}}. Only the first occurance should be kept.
}
\keyword{internal}
