% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{try_pattern_match}
\alias{try_pattern_match}
\title{Try to match a pattern against a value}
\usage{
try_pattern_match(value, pattern)
}
\arguments{
\item{value}{Character string to match}

\item{pattern}{Pattern definition list}
}
\value{
Tibble with results or NULL if no match
}
\description{
Try to match a pattern against a value
}
