% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{update_stat_patterns_to_pseudo}
\alias{update_stat_patterns_to_pseudo}
\title{Update statistical patterns to use pseudo-patterns}
\usage{
update_stat_patterns_to_pseudo()
}
\description{
Update statistical patterns to use pseudo-patterns
}
