% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{reorder_patterns_by_context}
\alias{reorder_patterns_by_context}
\title{Reorder patterns based on context}
\usage{
reorder_patterns_by_context(patterns, context_type)
}
\arguments{
\item{patterns}{List of patterns}

\item{context_type}{Context type from detect_context_type()}
}
\value{
Reordered patterns list
}
\description{
Reorder patterns based on context
}
