% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{rtf_to_ard}
\alias{rtf_to_ard}
\title{Convert an RTF Table into an ARD data frame}
\usage{
rtf_to_ard(file)
}
\arguments{
\item{file}{A string, the path to the input .rtf file.}
}
\value{
an R data frame following the ARD standard.
}
\description{
This function converts a table in RTF format into a data frame in R,
following the ARD standard. This is the top-level function of this package
which acts as a glue to the lower-level functions.
}
