% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{separate_indentation}
\alias{separate_indentation}
\title{Separate a single indented column into multiple columns}
\usage{
separate_indentation(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame where indented columns have been separated.
}
\description{
This function separates a single indented column into multiple columns inside
a data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}
}
\keyword{internal}
