% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{strip_pagination}
\alias{strip_pagination}
\title{Strip pagination from a table}
\usage{
strip_pagination(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame where pagination has been stripped.
}
\description{
This function combines the functionality of \code{\link[=strip_empty_rows]{strip_empty_rows()}},
\code{\link[=strip_header]{strip_header()}}, \code{\link[=strip_footer]{strip_footer()}} and \code{\link[=strip_repear_colnames]{strip_repear_colnames()}} into a
single function. If fine grained control is not required to remove certain
elements of a table, this function is preferred because the actions of
stripping table elements are dependent on each other (e.g., to remove
repeat column names, we first must identify these by stripping the header).
}
\keyword{internal}
