% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{strip_pagination}
\alias{strip_pagination}
\title{Remove pagination}
\usage{
strip_pagination(data)
}
\arguments{
\item{data}{A data frame generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}.}
}
\value{
A data frame with the pagination removed.
}
\description{
This function removes the rows representing pagination inside a data frame
generated by calling \code{\link[=rtf_to_html]{rtf_to_html()}}. These rows are signalled by repeated
column name headers. Only the first occurance should be kept.
}
\keyword{internal}
